Add files via upload
The main goal is to create a system that can detect motion within the camera's 
field of view, coupled with the ability to recognize faces. Such that we can handle 
unauthorized access to a confidential place and also provide safety from robbery, 
kidnapping, stalking, intrusions etc., with an automatic alert mechanism to provide a 
better solution for continuous monitoring and safeguarding our environment.